+++ src/main/kotlin/study/racinggame/race/DrawView.kt
@@ -1,0 +1,17 @@
+package study.racinggame.race
+
+import study.racinggame.domain.CarResponse
+
+class DrawView {
+
+    init {
+        println("실행 결과")
+    }
+
+    fun drawCarState(carResponses: List<CarResponse>) {
+        for (response in carResponses) {
+            println(response.getPositionRepresentation())
+        }
+        println()
+    }
+}
