+++ src/main/kotlin/study/racinggame/race/reader/ActualInputReader.kt
@@ -1,4 +1,4 @@
-package study.step3.race.reader
+package study.racinggame.race.reader
 
 class ActualInputReader : InputReader {
     override fun raedLine() = readlnOrNull()
