+++ src/test/kotlin/study/racinggame/domain/CarTest.kt
@@ -1,4 +1,4 @@
-package study.step3.domain
+package study.racinggame.domain
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
